Refactor logic out of kmq.ts
