+++ src/events/client/connect.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("connect");
 
-export default function connect(shardId: number) {
+export default function connectHandler(shardId: number) {
     logger.info(`Shard #${shardId} has connected.`)
 }
