+++ src/events/client/error.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("error");
-export default function error(err: Error, shardId: number) {
+export default function errorHandler(err: Error, shardId: number) {
     logger.error(`Shard #${shardId} encountered error: ${err.message}`);
 
 }
