+++ src/events/client/shardDisconnect.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("shardDisconnect");
 
-export default function shardDisconnect(err: Error, shardId: number){
+export default function shardDisconnectHandler(err: Error, shardId: number){
     logger.warn(`Shard #${shardId} disconnected. err = ${err.message}`);
 }
