+++ src/events/client/shardReady.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("shardReady");
 
-export default function shardReady(shardId: number){
+export default function shardReadyHandler(shardId: number){
     logger.info(`Shard #${shardId} ready.`);
 }
