+++ src/events/client/shardResume.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("shardResume");
 
-export default function shardResume(shardId: number){
+export default function shardResumeHandler(shardId: number){
     logger.info(`Shard #${shardId} resumed.`);
 }
