+++ src/events/client/voiceChannelLeave.ts
@@ -2,7 +2,7 @@ import * as Eris from "eris";
 import { state } from "../../kmq";
 import { checkBotIsAlone } from "../../helpers/discord_utils";
 
-export default async function voiceChannelLeave(member: Eris.Member, oldChannel: Eris.VoiceChannel) {
+export default async function voiceChannelLeaveHandler(member: Eris.Member, oldChannel: Eris.VoiceChannel) {
     const guildID = oldChannel.guild.id;
     const gameSession = state.gameSessions[guildID];
     await checkBotIsAlone(gameSession, oldChannel);
