+++ src/events/client/warn.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("warn");
 
-export default function warn(message: string, shardId: number){
+export default function warnHandler(message: string, shardId: number){
     logger.warn(`Shard #${shardId} encountered warning: ${message}`);
 }
