+++ src/events/process/uncaughtException.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("uncaughtException");
 
-export default function uncaughtException(err: Error) {
+export default function uncaughtExceptionHandler(err: Error) {
     logger.error(`Uncaught Exception. Reason: ${err}. Trace: ${err.stack}`);
 }
