+++ src/events/process/unhandledRejection.ts
@@ -1,6 +1,6 @@
 import _logger from "../../logger";
 const logger = _logger("unhandledRejection");
 
-export default function unhandledRejection(reason: Error, p: Promise<any>){
+export default function unhandledRejectionHandler(reason: Error, p: Promise<any>){
     logger.error(`Unhandled Rejection at: Promise ${p}. Reason: ${reason}. Trace: ${reason.stack}`);
 }
